Creada parte del perfil para la opción de carpetas, el recyclerView ya funciona
