+++ app/src/main/res/layout/activity_carpetas_view_holder.xml
@@ -26,6 +26,4 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toEndOf="@id/carpeta" />
 
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
