+++ app/src/main/res/layout/activity_perfil.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Perfil">
+    tools:context=".PerfilCarpetas">
 
     <ImageView
         android:id="@+id/imagen_perfil"
@@ -82,9 +82,9 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_carpetas"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
         android:layout_marginTop="10dp"
-        android:layout_marginBottom="10dp"
-        app:layout_constraintTop_toBottomOf="@id/icono_modificar_imagenes"/>
+        app:layout_constraintTop_toBottomOf="@id/icono_modificar_imagenes"
+        app:layout_constraintBottom_toBottomOf="parent"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
